[ANDROSDK-2353] remove BaseIdentifiableObAuVa.java merge BaseIdentifiableObjectKt.kt with BaseIdentifiableObject.kt and BaseNameableObjectKt.kt with BaseNameableObject.kt update files
