+++ core/src/main/java/org/hisp/dhis/android/core/event/EventFilter.kt
@@ -29,7 +29,7 @@
 package org.hisp.dhis.android.core.event
 
 import org.hisp.dhis.android.annotations.ModelBuilder
-import org.hisp.dhis.android.core.common.BaseIdentifiableObjectKt
+import org.hisp.dhis.android.core.common.BaseIdentifiableObject
 import org.hisp.dhis.android.core.common.CoreObject
 import java.util.Date
 
@@ -46,7 +46,7 @@ data class EventFilter(
     val programStage: String?,
     val description: String?,
     val eventQueryCriteria: EventQueryCriteria?,
-) : BaseIdentifiableObjectKt, CoreObject {
+) : BaseIdentifiableObject, CoreObject {
 
     fun program(): String? = program
     fun programStage(): String? = programStage
