+++ core/src/main/java/org/hisp/dhis/android/core/organisationunit/OrganisationUnitLevel.kt
@@ -29,7 +29,7 @@
 package org.hisp.dhis.android.core.organisationunit
 
 import org.hisp.dhis.android.annotations.ModelBuilder
-import org.hisp.dhis.android.core.common.BaseIdentifiableObjectKt
+import org.hisp.dhis.android.core.common.BaseIdentifiableObject
 import org.hisp.dhis.android.core.common.CoreObject
 import java.util.Date
 
@@ -43,7 +43,7 @@ data class OrganisationUnitLevel(
     override val lastUpdated: Date?,
     override val deleted: Boolean?,
     val level: Int?,
-) : BaseIdentifiableObjectKt, CoreObject {
+) : BaseIdentifiableObject, CoreObject {
 
     fun level(): Int? = level
