+++ core/src/main/java/org/hisp/dhis/android/core/program/ProgramRule.kt
@@ -29,7 +29,7 @@
 package org.hisp.dhis.android.core.program
 
 import org.hisp.dhis.android.annotations.ModelBuilder
-import org.hisp.dhis.android.core.common.BaseIdentifiableObjectKt
+import org.hisp.dhis.android.core.common.BaseIdentifiableObject
 import org.hisp.dhis.android.core.common.CoreObject
 import org.hisp.dhis.android.core.common.ObjectWithUid
 import java.util.Date
@@ -48,7 +48,7 @@ data class ProgramRule(
     val program: ObjectWithUid?,
     val programStage: ObjectWithUid?,
     val programRuleActions: List<ProgramRuleAction>?,
-) : BaseIdentifiableObjectKt, CoreObject {
+) : BaseIdentifiableObject, CoreObject {
 
     fun priority(): Int? = priority
     fun condition(): String? = condition
