+++ core/src/main/java/org/hisp/dhis/android/core/program/ProgramStageSection.kt
@@ -29,7 +29,7 @@
 package org.hisp.dhis.android.core.program
 
 import org.hisp.dhis.android.annotations.ModelBuilder
-import org.hisp.dhis.android.core.common.BaseIdentifiableObjectKt
+import org.hisp.dhis.android.core.common.BaseIdentifiableObject
 import org.hisp.dhis.android.core.common.CoreObject
 import org.hisp.dhis.android.core.common.ObjectWithUid
 import org.hisp.dhis.android.core.dataelement.DataElement
@@ -51,7 +51,7 @@ data class ProgramStageSection(
     val programStage: ObjectWithUid,
     val description: String?,
     val displayDescription: String?,
-) : BaseIdentifiableObjectKt, CoreObject {
+) : BaseIdentifiableObject, CoreObject {
 
     fun sortOrder(): Int? = sortOrder
     fun programIndicators(): List<ProgramIndicator>? = programIndicators
