+++ core/src/main/java/org/hisp/dhis/android/core/programstageworkinglist/ProgramStageWorkingList.kt
@@ -29,7 +29,7 @@
 package org.hisp.dhis.android.core.programstageworkinglist
 
 import org.hisp.dhis.android.annotations.ModelBuilder
-import org.hisp.dhis.android.core.common.BaseIdentifiableObjectKt
+import org.hisp.dhis.android.core.common.BaseIdentifiableObject
 import org.hisp.dhis.android.core.common.CoreObject
 import org.hisp.dhis.android.core.common.ObjectWithUid
 import java.util.Date
@@ -47,7 +47,7 @@ data class ProgramStageWorkingList(
     val program: ObjectWithUid,
     val programStage: ObjectWithUid,
     val programStageQueryCriteria: ProgramStageQueryCriteria?,
-) : BaseIdentifiableObjectKt, CoreObject {
+) : BaseIdentifiableObject, CoreObject {
 
     fun description(): String? = description
     fun program(): ObjectWithUid = program
